PlotGroups(maSigPro)	R Documentation

Function for plotting gene expression profile at different experimental groups

Description

This function displays the gene expression profile for each experimental group in a time series gene expression experiment.

Usage


PlotGroups(data, edesign = NULL, time = edesign[,1], groups = edesign[,c(3:ncol(edesign))], 
           repvect = edesign[,2], show.fit = FALSE, dis = NULL, step.method = "backward", 
           min.obs = 2, alfa = 0.01, show.lines = TRUE, groups.vector = NULL, xlab = "time", cex.xaxis = 1, 
           ylim = NULL, main = NULL, cexlab = 0.8, legend = TRUE, sub = NULL)

Arguments

data 	 vector or matrix containing the gene expression data

edesign 	 matrix describing experimental design. Rows must be arrays and columns experiment descriptors

time 	 vector indicating time assigment for each array

groups 	 matrix indicating experimental group to which each array is assigned

repvect 	 index vector indicating experimental replicates

show.fit 	 logical indicating whether regression fit curves must be plotted

dis 	 regression design matrix

step.method 	 stepwise regression method to fit models for cluster mean profiles. It can be either "backward", "forward", "two.ways.backward" or "two.ways.forward"

min.obs 	 minimal number of observations for a gene to be included in the analysis

alfa 	 significance level used for variable selection in the stepwise regression

show.lines 	 logical indicating whether a line must be drawn joining plotted data points for reach group

groups.vector 	 vector indicating experimental group to which each variable belongs

xlab 	 label for the x axis

cex.xaxis 	 graphical parameter maginfication to be used for x axis in plotting functions

ylim 	 range of the y axis

main 	 plot main title

cexlab 	 graphical parameter maginfication to be used for x axis label in plotting functions

legend 	 logical indicating whether legend must be added when plotting profiles

sub 	 plot subtitle

Details

To compute experimental groups either a edesign object must be provided, or separate values must be given for the time, repvect and groups arguments. newline When data is a matrix, the average expression value is displayed. newline When there are array replicates in the data (as indicated by repvect), values are averaged by repvect. newline When show.fit is TRUE the stepwise regression fit for the data will be computed and the regression curves will be displayed. If data is a matrix of genes, the regression fit will be computed for the averaged expression value.

Value

Plot of gene expression profiles by-group.

Author(s)

Ana Conesa, aconesa@ivia.es; María José Nueda, mj.nueda@ua.es

References

Conesa, A., Nueda M.J., Alberto Ferrer, A., Talón, T. 2005. maSigPro: a Method to Identify Significant Differential Expression Profiles in Time-Course Microarray Experiments.

See Also

PlotProfiles

Examples



#### GENERATE TIME COURSE DATA
## generate n random gene expression profiles of a data set with 
## one control plus 3 treatments, 3 time points and r replicates per time point.

tc.GENE <- function(n, r,
             var11 = 0.01, var12 = 0.01,var13 = 0.01,
             var21 = 0.01, var22 = 0.01, var23 =0.01,
             var31 = 0.01, var32 = 0.01, var33 = 0.01,
             var41 = 0.01, var42 = 0.01, var43 = 0.01,
             a1 = 0, a2 = 0, a3 = 0, a4 = 0,
             b1 = 0, b2 = 0, b3 = 0, b4 = 0,
             c1 = 0, c2 = 0, c3 = 0, c4 = 0)
{

  tc.dat <- NULL
  for (i in 1:n) {
    Ctl <- c(rnorm(r, a1, var11), rnorm(r, b1, var12), rnorm(r, c1, var13))  # Ctl group
    Tr1 <- c(rnorm(r, a2, var21), rnorm(r, b2, var22), rnorm(r, c2, var23))  # Tr1 group
    Tr2 <- c(rnorm(r, a3, var31), rnorm(r, b3, var32), rnorm(r, c3, var33))  # Tr2 group
    Tr3 <- c(rnorm(r, a4, var41), rnorm(r, b4, var42), rnorm(r, c4, var43))  # Tr3 group
    gene <- c(Ctl, Tr1, Tr2, Tr3)
    tc.dat <- rbind(tc.dat, gene)
  }
  tc.dat
}

## create 10 genes with profile differences between Ctl, Tr2, and Tr3 groups
tc.DATA <- tc.GENE(n = 10,r = 3, b3 = 0.8, c3 = -1, a4 = -0.1, b4 = -0.8, c4 = -1.2)
rownames(tc.DATA) <- paste("gene", c(1:10), sep = "")
colnames(tc.DATA) <- paste("Array", c(1:36), sep = "")

#### CREATE EXPERIMENTAL DESIGN
Time <- rep(c(rep(c(1:3), each = 3)), 4)
Replicates <- rep(c(1:12), each = 3)
Ctl <- c(rep(1, 9), rep(0, 27))
Tr1 <- c(rep(0, 9), rep(1, 9), rep(0, 18))
Tr2 <- c(rep(0, 18), rep(1, 9), rep(0, 9))
Tr3 <- c(rep(0, 27), rep(1, 9))

PlotGroups (tc.DATA, time = Time, repvect = Replicates, groups = cbind(Ctl, Tr1, Tr2, Tr3))


[Package maSigPro version 1.0 Index]