data.abiotic(maSigPro)	R Documentation

Gene expression data potato abiotic stress

Description

data.abiotic contains gene expression of a time course microarray experiment where potato plants were submitted to 3 different abiotic stresses.

Usage

data(data.abiotic)

Format

A data frame with 1000 observations on the following 36 variables.

Control_3H_1

a numeric vector

Control_3H_2

a numeric vector

Control_3H_3

a numeric vector

Control_9H_1

a numeric vector

Control_9H_2

a numeric vector

Control_9H_3

a numeric vector

Control_27H_1

a numeric vector

Control_27H_2

a numeric vector

Control_27H_3

a numeric vector

Cold_3H_1

a numeric vector

Cold_3H_2

a numeric vector

Cold_3H_3

a numeric vector

Cold_9H_1

a numeric vector

Cold_9H_2

a numeric vector

Cold_9H_3

a numeric vector

Cold_27H_1

a numeric vector

Cold_27H_2

a numeric vector

Cold_27H_3

a numeric vector

Heat_3H_1

a numeric vector

Heat_3H_2

a numeric vector

Heat_3H_3

a numeric vector

Heat_9H_1

a numeric vector

Heat_9H_2

a numeric vector

Heat_9H_3

a numeric vector

Heat_27H_1

a numeric vector

Heat_27H_2

a numeric vector

Heat_27H_3

a numeric vector

Salt_3H_1

a numeric vector

Salt_3H_2

a numeric vector

Salt_3H_3

a numeric vector

Salt_9H_1

a numeric vector

Salt_9H_2

a numeric vector

Salt_9H_3

a numeric vector

Salt_27H_1

a numeric vector

Salt_27H_2

a numeric vector

Salt_27H_3

a numeric vector

Details

This data set is part of a larger experiment in wich gene expression was monitored in both roots and leaves using a 11K cDNA potato chip. This example data set contains a ramdom subset of 1000 genes of the leave study.

References

Rensink WA, Iobst S, Hart A, Stegalkina S, Liu J, Buell CR. Gene expression profiling of potato responses to cold, heat, and salt stress. Funct Integr Genomics. 2005 Apr 22.

Examples


data(data.abiotic)


[Package maSigPro version 1.0 Index]