edesign.abiotic(maSigPro)	R Documentation

Experimental design potato abiotic stress

Description

edesign.abiotic contains experimental set up of a time course microarray experiment where potato plants were submitted to 3 different abiotic stresses.

Usage

data(edesign.abiotic)

Format

A matrix with 36 rows and 6 columns
rows [1:36] "Control 3h 1" "Control 3h 2" "Control 3h 3" "Control 9h 1" ... newline columns [1:6] "Time" "Replicates" "Control" "Cold" "Heat" "Salt"

Details

Arrays are given in rows and experiment descriptors are given in columns. Row names contain array names. newline "Time" indicates the values that variable Time takes in each hybridization. newline "Replicates" is an index indicating replicate hyridizations, i.e. hybridizations are numbered, giving replicates the same number. newline "Control", "Cold", "Heat" and "Salt" columns indicate array assigment to experimental groups, coding with 1 and 0 whether each array belongs to that group or not.

References

Rensink WA, Iobst S, Hart A, Stegalkina S, Liu J, Buell CR. Gene expression profiling of potato responses to cold, heat, and salt stress. Funct Integr Genomics. 2005 Apr 22.

Examples


data(edesignCR)

[Package maSigPro version 1.0 Index]