edesignCT(maSigPro)	R Documentation

Experimental design with a shared time

Description

edesignCT contains the experimental set up of a time course microarray experiment where there is a common starting point for the different experimental groups.

Usage

data(edesignCT)

Format

A matrix with 32 rows and 7 colums
rows [1:32] "Array1" "Array2" "Array3" "Array4" ... newline columns [1:7] "Time" "Replicates" "Control" "Tissue1" "Tissue2" "Tissue3" "Tissue4"

Details

Arrays are given in rows and experiment descriptors are given in columns. Row names contain array names. newline "Time" indicates the values that variable Time takes in each hybridization. There are 4 time points, which allows an up to 3 degree regression polynome. newline "Replicates" is an index indicating replicate hyridizations, i.e. hybridizations are numbered, giving replicates the same number. newline "Control", "Tissue1", "Tissue2", "Tissue3" and "Tissue4" columns indicate array assigment to experimental groups, coding with 1 and 0 whether each array belongs to that group or not.

Examples


data(edesignCT)


[Package maSigPro version 1.0 Index]