make.design.matrix(maSigPro)	R Documentation

Make a design matrix for regression fit of time series gene expression experiments

Description

make.design.matrix creates the design matrix of dummies for fitting time series micorarray gene expression experiments.

Usage


make.design.matrix(edesign, degree = 2, time.col = 1, 
                   repl.col = 2, group.cols = c(3:ncol(edesign)))

Arguments

edesign 	 matrix describing experimental design. Rows must be arrays and columns experiment descriptors

degree 	 the degree of the regression fit polynome. degree = 1 returns linear regression, degree = 2 returns quadratic regression, etc

time.col 	 column number in edesign containing time values. Default is first column

repl.col 	 column number in edesign containing coding for replicate arrays. Default is second column

group.cols 	 column numbers in edesign indicating the coding for each experimental group (treatment, tissue, ...). See details

Details

rownames of edesign object should contain the arrays naming (i.e. array1, array2, ...). colnames of edesign must contain the names of experiment descriptors(i.e. "Time", "Replicates", "Treatment A", "Treatment B", etc.). for each experimental group a different column must be present in edesign, coding with 1 and 0 whether each array belongs to that group or not.

make.design.matrix returns a design matrix where rows represent arrays and column variables of time, dummies and their interactions for up to the degree given. Dummies show the relative effect of each experimental group related to the first one. Single dummies indicate the abcissa component of each group. $Time*dummy$ variables indicate slope changes, $Time^2*dummy$ indicates curvature changes. Higher grade values could model complex responses. In case experimental groups share a initial state (i.e. common time 0), no single dummies are modeled.

Value

dis 	 design matrix of dummies for fitting time series

groups.vector 	 vector coding the experimental group to which each variable belongs to

edesign 	 edesign value passed as argument

Author(s)

Ana Conesa, aconesa@ivia.es; María José Nueda, mj.nueda@ua.es

References

Conesa, A., Nueda M.J., Alberto Ferrer, A., Talón, T. 2005. maSigPro: a Method to Identify Significant Differential Expression Profiles in Time-Course Microarray Experiments.

Examples


data(edesign.abiotic, edesignCT)
make.design.matrix(edesign.abiotic)  # quadratic model
make.design.matrix(edesignCT, degree = 3)  # cubic model with common starting time point

[Package maSigPro version 1.0 Index]