reg.coeffs(maSigPro)	R Documentation

Calculate true variables regression coefficients

Description

reg.coeffs calculates back regression coefficients for true variables (experimental groups) from dummy variables regression coefficients.

Usage


reg.coeffs(coefficients, indepen = groups.vector[1], groups.vector, group)

Arguments

coefficients 	 vector of regression coefficients obtained from a regression model with dummy variables

indepen 	 idependent variable of the regression formula

groups.vector 	 vector indicating the true variable of each variable in coefficients

group 	 true variable for which regression coefficients are to be computed

Details

regression coefficients in coefficients vector should be ordered by polynomial degree in a regression formula, ie: intercept, $x$ term, $x^2$ term, $x^3$ term, and so on...

Value

reg.coeff 	 vector of calculated regression coefficients

Author(s)

Ana Conesa, aconesa@ivia.es; María José Nueda, mj.nueda@ua.es

References

Conesa, A., Nueda M.J., Alberto Ferrer, A., Talón, T. 2005. maSigPro: a Method to Identify Significant Differential Expression Profiles in Time-Course Microarray Experiments.

Examples


groups.vector <-c("CT", "T1vsCT", "T2vsCT", "T1vsCT","T2vsCT", "CT", "T1vsCT", "T2vsCT")
coefficients <- c(0.1, 1.2, -0.8, 3.3, 0.4, 0.0, 2.1, -0.9)
## calculate true regression coefficients for variable "T1"
reg.coeffs(coefficients, groups.vector = groups.vector, group = "T1")

[Package maSigPro version 1.0 Index]