see.genes(maSigPro)	R Documentation

Wrapping function for visualization of gene expression values of time course experiments

Description

This function provides visualisation tools for gene expression values in a time course experiment. The function first calls the heatmap function for a general overview of experiment results. Next a partioning of the data is generated using a clustering method. The results of the clustering are visualized both as gene expression profiles extended along all arrays in the experiment, as provided by the plot.profiles function, and as summary expression profiles for comparison among experimental groups.

Usage


see.genes(data, edesign = data$edesign, time.col = 1, repl.col = 2, group.cols = 
          c(3:ncol(edesign)), names.groups = colnames(edesign)[3:ncol(edesign)], 
          cluster.data = 1, groups.vector = data$groups.vector, k = 9, 
          cluster.method = "kmeans", distance = "cor", agglo.method = "complete", 
          show.fit = FALSE, dis = NULL, step.method = "backward", min.obs = 3, 
          alfa = 0.01, show.lines = TRUE, iter.max = 500, cexlab = 1, legend = TRUE, 
          newX11 = TRUE, ylim = NULL, main = NULL, ...)

Arguments

data 	 either matrix or a list containing the gene expression data, typically a get.siggenes object

edesign 	 matrix of experimental design

time.col 	 column in edesign containing time values. Default is first column

repl.col 	 column in edesign containing coding for replicates arrays. Default is second column

group.cols 	 columns indicating the coding for each group (treatment, tissue,...) in the experiment (see details)

names.groups 	 names for experimental groups

cluster.data 	 type of data used by the cluster algorithm (see details)

groups.vector 	 Vector indicating the experimental group to which each variable belongs

k 	 number of clusters for data partioning

cluster.method 	 clustering method for data partioning

distance 	 distance measurement function for when cluster.method is "hclust"

agglo.method 	 aggregation method used when cluster.method is "hclust"

show.fit 	 logical indicating whether regression fit curves must be plotted

dis 	 regression design matrix

step.method 	 stepwise regression method to fit models for cluster mean profiles. Can be either "backward", "forward", "two.ways.backward" or "two.ways.forward"

min.obs 	 minimal number of observations for a gene to be included in the analysis

alfa 	 significance level used for variable selection in the stepwise regression

show.lines 	 logical indicating whether a line must be drawn joining plotted data points for reach group

iter.max 	 max. number of iteration when cluster.method = "kmeans"

cexlab 	 graphical parameter maginfication to be used for x labels in plotting functions

legend 	 logical indicating whether legend must be added when plotting profiles

main 	 plot title

ylim 	 range of the y axis to be used by PlotProfiles and PlotGroups

newX11 	 If TRUE, plot each type of plot in a diferent graphical device

... 	 other graphical function argument

Details

Data can be provided either as a single data matrix of expression values, or a get.siggenes object. In the later case the other argument of the fuction can be taken directly from data.

Data clustering can be done on the basis of either the original expression values, the regression coefficients, or the t.scores. In case data is a get.siggenes object, this is given by providing the element names of the list c("sig.profiles","coefficients","t.score") of their list position (1,2 or 3).

Value

Experiment wide gene profiles and by group profiles plots are generated for each data cluster in the graphical device.

cut 	 vector indicating gene partioning into clusters

c.algo.used 	 clustering algorith used for data partioning

groups 	 groups matrix used for plotting functions

Author(s)

Ana Conesa, aconesa@ivia.es; María José Nueda, mj.nueda@ua.es

References

Conesa, A., Nueda M.J., Alberto Ferrer, A., Talón, T. 2005. maSigPro: a Method to Identify Significant Differential Expression Profiles in Time-Course Microarray Experiments.

See Also

PlotProfiles, PlotGroups

Examples



#### GENERATE TIME COURSE DATA
## generate n random gene expression profiles of a data set with 
## one control plus 3 treatments, 3 time points and r replicates per time point.

tc.GENE <- function(n, r,
             var11 = 0.01, var12 = 0.01,var13 = 0.01,
             var21 = 0.01, var22 = 0.01, var23 =0.01,
             var31 = 0.01, var32 = 0.01, var33 = 0.01,
             var41 = 0.01, var42 = 0.01, var43 = 0.01,
             a1 = 0, a2 = 0, a3 = 0, a4 = 0,
             b1 = 0, b2 = 0, b3 = 0, b4 = 0,
             c1 = 0, c2 = 0, c3 = 0, c4 = 0)
{

  tc.dat <- NULL
  for (i in 1:n) {
    Ctl <- c(rnorm(r, a1, var11), rnorm(r, b1, var12), rnorm(r, c1, var13))  # Ctl group
    Tr1 <- c(rnorm(r, a2, var21), rnorm(r, b2, var22), rnorm(r, c2, var23))  # Tr1 group
    Tr2 <- c(rnorm(r, a3, var31), rnorm(r, b3, var32), rnorm(r, c3, var33))  # Tr2 group
    Tr3 <- c(rnorm(r, a4, var41), rnorm(r, b4, var42), rnorm(r, c4, var43))  # Tr3 group
    gene <- c(Ctl, Tr1, Tr2, Tr3)
    tc.dat <- rbind(tc.dat, gene)
  }
  tc.dat
}

## Create 270 flat profiles
flat <- tc.GENE(n = 270, r = 3)
## Create 10 genes with profile differences between Ctl and Tr1 groups
twodiff <- tc.GENE (n = 10, r = 3, b2 = 0.5, c2 = 1.3)
## Create 10 genes with profile differences between Ctl, Tr2, and Tr3 groups
threediff <- tc.GENE(n = 10, r = 3, b3 = 0.8, c3 = -1, a4 = -0.1, b4 = -0.8, c4 = -1.2)
## Create 10 genes with profile differences between Ctl and Tr2 and different variance
vardiff <- tc.GENE(n = 10, r = 3, a3 = 0.7, b3 = 1, c3 = 1.2, var32 = 0.03, var33 = 0.03)
## Create dataset
tc.DATA <- rbind(flat, twodiff, threediff, vardiff)
rownames(tc.DATA) <- paste("feature", c(1:300), sep = "")
colnames(tc.DATA) <- paste("Array", c(1:36), sep = "")
tc.DATA [sample(c(1:(300*36)), 300)] <- NA  # introduce missing values

#### CREATE EXPERIMENTAL DESIGN
Time <- rep(c(rep(c(1:3), each = 3)), 4)
Replicates <- rep(c(1:12), each = 3)
Control <- c(rep(1, 9), rep(0, 27))
Treat1 <- c(rep(0, 9), rep(1, 9), rep(0, 18))
Treat2 <- c(rep(0, 18), rep(1, 9), rep(0,9))
Treat3 <- c(rep(0, 27), rep(1, 9))
edesign <- cbind(Time, Replicates, Control, Treat1, Treat2, Treat3)
rownames(edesign) <- paste("Array", c(1:36), sep = "")

see.genes(tc.DATA, edesign = edesign, k = 4, main = "Time Course")

# This will show the regression fit curve
dise <- make.design.matrix(edesign)
see.genes(tc.DATA, edesign = edesign, k = 4, main = "Time Course", show.fit = TRUE, 
          dis = dise$dis, groups.vector = dise$groups.vector, distance = "euclidean")


[Package maSigPro version 1.0 Index]