suma2Venn(maSigPro)	R Documentation

Creates a Venn Diagram from a matrix of characters

Description

suma2Venn transforms a matrix of characters into a binary matrix and creates a vennDiagram of the common elements between columns

Usage


suma2Venn(x, ...)

Arguments

x 	 data frame of character values

... 	 plotting arguments for the vennDiagram function

Details

suma2Venn creates a list of all elements of a matrix or data frame of characters and computes the presence/absence of each element in each column of the matrix. This results is a numeric matrix of 1 and 0 which can be taken by the vennDiagram to generate a Venn Plot

Value

suma2Venn returns a Venn Plot such as that created by the vennDiagram funcion

Author(s)

Ana Conesa, aconesa@ivia.es

See Also

vennDiagram

Examples


a <- c("a","b","c", "d", "e", NA, NA)
b <- c("a","b","f", NA, NA, NA, NA)
c <- c("b","e","f", "h", "i", "j", "k")
x <- cbind(a, b,c)
suma2Venn(x)


[Package maSigPro version 1.0 Index]